pyproject.toml
[project]
name = "cash-flow-commander"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "datetime>=5.5",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "numpy>=2.3.0",
    "pandas>=2.3.0",
    "psycopg2-binary>=2.9.10",
    "pygsheets>=2.0.6",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "tabulate>=0.9.0",
    "tqdm>=4.67.1",
]

[tool.flake8]
max-line-length = 120
max-complexity = 15
extend-ignore = ["E203", "W503"]

[tool.mypy]
ignore_missing_imports = true
show_error_codes = true
